requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
future==0.18.2
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
pytesseract==0.3.0
pytz==2020.1
sqlparse==0.3.1
tkintertable==1.3.2
virtualenv==16.7.5
